requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
sqlalchemy
pymupdf
pdfplumber
python-dotenv
